Revert "refactor: partial refactoring and compliance with new rules  (#349)"
This reverts commit 57280ffebcab52256588188f28f71c56a829883d.

+++ system/library/games/cs2/action.php
@@ -16,16 +16,11 @@
  * limitations under the License.
  */
 
-use EngineGP\System;
-use EngineGP\Model\Game;
-use EngineGP\Model\Parameters;
-use EngineGP\Infrastructure\RemoteAccess\SshClient;
-use EngineGP\Infrastructure\RemoteAccess\SftpClient;
-use EngineGP\Infrastructure\Network\InternalIpFetcher;
-
 if (!defined('EGP')) {
     exit(header('Refresh: 0; URL=http://' . $_SERVER['HTTP_HOST'] . '/404'));
 }
+
+include(LIB . 'games/actions.php');
 
 class action extends actions
 {
@@ -39,31 +34,34 @@ class action extends actions
         $sql->query('SELECT `install` FROM `tarifs` WHERE `id`="' . $server['tarif'] . '" LIMIT 1');
         $tarif = $sql->get();
 
+        include(LIB . 'ssh.php');
+
         $sql->query('SELECT `address`, `passwd` FROM `units` WHERE `id`="' . $server['unit'] . '" LIMIT 1');
         $unit = $sql->get();
 
-        $sshClient = new SshClient($unit['address'], 'root', $unit['passwd']);
-        $sftpClient = new SftpClient($unit['address'], 'root', $unit['passwd']);
-        $internalIpFetcher = new InternalIpFetcher($sshClient);
+        // Проверка ssh соедниения пу с локацией
+        if (!$ssh->auth($unit['passwd'], $unit['address'])) {
+            return ['e' => sys::text('error', 'ssh')];
+        }
 
-        $ip = $internalIpFetcher->getInternalIp();
+        $ip = $ssh->getInternalIp();
         $port = $server['port'];
         $server_address = $server['address'] . ':' . $server['port'];
 
-        $serverSystemdStatus = trim($sshClient->execute('sudo systemctl show -p ActiveState server' . $server['uid'] . '.scope | awk -F \'=\' \'{print $2}\'', false));
+        $serverSystemdStatus = trim($ssh->get('sudo systemctl show -p ActiveState server' . $server['uid'] . '.scope | awk -F \'=\' \'{print $2}\''));
 
         if ($serverSystemdStatus == 'failed') {
-            $sshClient->execute('sudo systemctl stop server' . $server['uid'] . '.scope');
-            $sshClient->execute('sudo systemctl reset-failed server' . $server['uid'] . '.scope');
+            $ssh->set('sudo systemctl stop server' . $server['uid'] . '.scope');
+            $ssh->set('sudo systemctl reset-failed server' . $server['uid'] . '.scope');
         }
 
         // Убить процессы
-        $sshClient->execute('kill -9 `ps aux | grep s_' . $server['uid'] . ' | grep -v grep | awk ' . "'{print $2}'" . ' | xargs;'
+        $ssh->set('kill -9 `ps aux | grep s_' . $server['uid'] . ' | grep -v grep | awk ' . "'{print $2}'" . ' | xargs;'
             . 'lsof -i@' . $server_address . ' | awk ' . "'{print $2}'" . ' | grep -v PID | xargs`; sudo -u server' . $server['uid'] . ' tmux kill-session -t server' . $server['uid']);
 
         // Проверка наличия .steam
         $checkLinkCommand = 'ls -la' . $tarif['install'] . $server['uid'];
-        $checkLinkOutput = $sshClient->execute($checkLinkCommand, false);
+        $checkLinkOutput = $ssh->get($checkLinkCommand);
 
         // Если .steam отсуствует, создаём каталог и символическую ссылку на steamclient.so
         if (strpos($checkLinkOutput, '.steam') === false) {
@@ -71,16 +69,18 @@ class action extends actions
                 . 'ln -s ' . $cfg['steamcmd'] . '/linux64/steamclient.so ' . $tarif['install'] . $server['uid'] . '/.steam/sdk64/' . ';'
                 . 'chown -R server' . $server['uid'] . ':servers ' . $tarif['install'] . $server['uid'] . '/.steam' . ';'
                 . 'find ' . $tarif['install'] . $server['uid'] . '/.steam' . ' -type d -exec chmod 700 {} \;';
-            $sshClient->execute($createLinkCommand);
+            $ssh->get($createLinkCommand);
         }
 
-        $sshClient->execute('chmod +x ' . $tarif['install'] . $server['uid'] . '/game/bin/linuxsteamrt64/cs2');
+        $ssh->get('chmod +x ' . $tarif['install'] . $server['uid'] . '/game/bin/linuxsteamrt64/cs2');
 
         // Проверка наличия стартовой карты
-        $output = $sshClient->execute('cd ' . $tarif['install'] . $server['uid'] . '/game/csgo/maps/ && du -ah | grep -e "\.vpk$" | awk \'{print $2}\'', false);
+        $ssh->set('cd ' . $tarif['install'] . $server['uid'] . '/game/csgo/maps/ && du -ah | grep -e "\.vpk$" | awk \'{print $2}\'');
 
-        if (Game::map($server['map_start'], $output)) {
-            return ['e' => System::updtext(System::text('servers', 'nomap'), ['map' => $server['map_start'] . '.vpk'])];
+        include_once(LIB . 'games/games.php');
+
+        if (games::map($server['map_start'], $ssh->get())) {
+            return ['e' => sys::updtext(sys::text('servers', 'nomap'), ['map' => $server['map_start'] . '.vpk'])];
         }
 
         // Античит VAC
@@ -112,14 +112,14 @@ class action extends actions
         $bash = './game/bin/linuxsteamrt64/cs2 -dedicated -debug -norestart -condebug console.log -usercon -ip ' . $ip . ' -port ' . $port . ' -maxplayers ' . $server['slots_start'] . ' -tickrate ' . $server['tickrate'] . ' ' . $mod . ' +servercfgfile server.cfg ' . $map . ' ' . $vac . ' ' . $bots . ' ' . $tv;
 
         // Временный файл
-        $temp = System::temp($bash);
+        $temp = sys::temp($bash);
 
         // Обновление файла start.sh
-        $sftpClient->putFile($temp, $tarif['install'] . $server['uid'] . '/start.sh');
-        $sshClient->execute('chmod 0500' . ' ' . $tarif['install'] . $server['uid'] . '/start.sh');
+        $ssh->setfile($temp, $tarif['install'] . $server['uid'] . '/start.sh');
+        $ssh->set('chmod 0500' . ' ' . $tarif['install'] . $server['uid'] . '/start.sh');
 
         // Строка запуска
-        $sshClient->execute('cd ' . $tarif['install'] . $server['uid'] . ';' // переход в директорию игрового сервера
+        $ssh->set('cd ' . $tarif['install'] . $server['uid'] . ';' // переход в директорию игрового сервера
             . 'rm *.pid;' // Удаление *.pid файлов
             . 'sudo -u server' . $server['uid'] . ' mkdir -p game/csgo/oldstart;' // Создание папки логов
             . 'cat game/csgo/console.log >> game/csgo/oldstart/' . date('d.m.Y_H:i:s', $server['time_start']) . '.log; rm game/csgo/console.log; rm game/csgo/oldstart/01.01.1970_03:00:00.log;'  // Перемещение лога предыдущего запуска
@@ -134,11 +134,8 @@ class action extends actions
         // Сброс кеша
         actions::clmcache($id);
 
-        System::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0, 'players' => '']);
-        System::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0]);
-
-        $sshClient->disconnect();
-        $sftpClient->disconnect();
+        sys::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0, 'players' => '']);
+        sys::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => $type, 'online' => 0]);
 
         return ['s' => 'ok'];
     }
@@ -152,6 +149,10 @@ class action extends actions
             return ['maps' => $mcache->get('server_maps_change_' . $id)];
         }
 
+        include(LIB . 'ssh.php');
+
+        include(LIB . 'games/games.php');
+
         $sql->query('SELECT `uid`, `unit`, `game`, `tarif`, `online`, `players`, `name` FROM `servers` WHERE `id`="' . $id . '" LIMIT 1');
         $server = $sql->get();
 
@@ -161,10 +162,13 @@ class action extends actions
         $sql->query('SELECT `install` FROM `tarifs` WHERE `id`="' . $server['tarif'] . '" LIMIT 1');
         $tarif = $sql->get();
 
-        $sshClient = new SshClient($unit['address'], 'root', $unit['passwd']);
+        // Проверка ssh соедниения пу с локацией
+        if (!$ssh->auth($unit['passwd'], $unit['address'])) {
+            return ['e' => sys::text('error', 'ssh')];
+        }
 
         // Массив карт игрового сервера (папка "maps")
-        $aMaps = explode("\n", $sshClient->execute('cd ' . $tarif['install'] . $server['uid'] . '/game/csgo/maps/ && du -ah | grep -e "\.vpk$" | awk \'{print $2}\'', false));
+        $aMaps = explode("\n", $ssh->get('cd ' . $tarif['install'] . $server['uid'] . '/game/csgo/maps/ && du -ah | grep -e "\.vpk$" | awk \'{print $2}\''));
 
         // Удаление пустого элемента
         unset($aMaps[count($aMaps) - 1]);
@@ -177,12 +181,12 @@ class action extends actions
             $map = str_replace('|', '/', $map);
 
             // Проверка наличия выбранной карты
-            if (Game::map($map, $aMaps)) {
-                return ['e' => System::updtext(System::text('servers', 'change'), ['map' => $map . '.vpk'])];
+            if (games::map($map, $aMaps)) {
+                return ['e' => sys::updtext(sys::text('servers', 'change'), ['map' => $map . '.vpk'])];
             }
 
             // Отправка команды changelevel
-            $sshClient->execute('sudo -u server' . $server['uid'] . ' tmux send-keys -t s_' . $server['uid'] . ' "changelevel ' . System::cmd($map) . '" C-m');
+            $ssh->set('sudo -u server' . $server['uid'] . ' tmux send-keys -t s_' . $server['uid'] . ' "changelevel ' . sys::cmd($map) . '" C-m');
 
             // Обновление информации в базе
             $sql->query('UPDATE `servers` set `status`="change" WHERE `id`="' . $id . '" LIMIT 1');
@@ -190,10 +194,8 @@ class action extends actions
             // Сброс кеша
             actions::clmcache($id);
 
-            System::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'change', 'online' => $server['online'], 'players' => base64_decode($server['players'])]);
-            System::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'change', 'online' => $server['online']]);
-
-            $sshClient->disconnect();
+            sys::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'change', 'online' => $server['online'], 'players' => base64_decode($server['players'])]);
+            sys::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'change', 'online' => $server['online']]);
 
             return ['s' => 'ok'];
         }
@@ -233,6 +235,8 @@ class action extends actions
     {
         global $cfg, $sql, $user, $start_point;
 
+        include(LIB . 'ssh.php');
+
         $sql->query('SELECT `uid`, `unit`, `tarif`, `game`, `name`, `ftp`, `update` FROM `servers` WHERE `id`="' . $id . '" LIMIT 1');
         $server = $sql->get();
 
@@ -240,7 +244,7 @@ class action extends actions
         $update = $server['update'] + $cfg['update'][$server['game']] * 60;
 
         if ($update > $start_point and $user['group'] != 'admin') {
-            return ['e' => System::updtext(System::text('servers', 'update'), ['time' => System::date('max', $update)])];
+            return ['e' => sys::updtext(sys::text('servers', 'update'), ['time' => sys::date('max', $update)])];
         }
 
         $sql->query('SELECT `address`, `passwd`, `sql_login`, `sql_passwd`, `sql_port`, `sql_ftp` FROM `units` WHERE `id`="' . $server['unit'] . '" LIMIT 1');
@@ -249,32 +253,33 @@ class action extends actions
         $sql->query('SELECT `install`, `plugins_install` FROM `tarifs` WHERE `id`="' . $server['tarif'] . '" LIMIT 1');
         $tarif = $sql->get();
 
-        $sshClient = new SshClient($unit['address'], 'root', $unit['passwd']);
+        // Проверка ssh соедниения пу с локацией
+        if (!$ssh->auth($unit['passwd'], $unit['address'])) {
+            return ['e' => sys::text('error', 'ssh')];
+        }
 
         // Директория игрового сервера
         $install = $tarif['install'] . $server['uid'];
 
-        $sshClient->execute('cd ' . $cfg['steamcmd'] . ' && ' . 'tmux new-session -ds u_' . $server['uid'] . ' sh -c "'
+        $ssh->set('cd ' . $cfg['steamcmd'] . ' && ' . 'tmux new-session -ds u_' . $server['uid'] . ' sh -c "'
             . './steamcmd.sh +login anonymous +force_install_dir "' . $install . '" +app_update 730 +quit;'
             . 'cd ' . $install . ';'
             . 'chown -R server' . $server['uid'] . ':servers .;'
             . 'find . -type d -exec chmod 700 {} \;;'
             . 'find . -type f -exec chmod 600 {} \;;'
-            . 'chmod 500 ' . Parameters::$aFileGame[$server['game']] . '"');
+            . 'chmod 500 ' . params::$aFileGame[$server['game']] . '"');
 
         // Обновление информации в базе
         $sql->query('UPDATE `servers` set `status`="update", `update`="' . $start_point . '" WHERE `id`="' . $id . '" LIMIT 1');
 
         // Логирование
-        $sql->query('INSERT INTO `logs_sys` set `user`="' . $user['id'] . '", `server`="' . $id . '", `text`="' . System::text('syslogs', 'update') . '", `time`="' . $start_point . '"');
+        $sql->query('INSERT INTO `logs_sys` set `user`="' . $user['id'] . '", `server`="' . $id . '", `text`="' . sys::text('syslogs', 'update') . '", `time`="' . $start_point . '"');
 
         // Сброс кеша
         actions::clmcache($id);
 
-        System::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'update', 'online' => 0, 'players' => '']);
-        System::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'update', 'online' => 0]);
-
-        $sshClient->disconnect();
+        sys::reset_mcache('server_scan_mon_pl_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'update', 'online' => 0, 'players' => '']);
+        sys::reset_mcache('server_scan_mon_' . $id, $id, ['name' => $server['name'], 'game' => $server['game'], 'status' => 'update', 'online' => 0]);
 
         return ['s' => 'ok'];
     }
